+++ footer.php
@@ -1,1 +1,10 @@
 <? do_action('basset/footer') ?>
+
+<footer>
+<?php
+// Standard Footer Components
+do_action('basset/social_icons');
+do_action('basset/footer_utility_nav');
+do_action('basset/footnotes');
+?>
+</footer>